removing old components
