--- src/View/Personalizations/Components/Select/Searchable.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace TallStackUi\View\Personalizations\Components\Select;
-
-use TallStackUi\View\Components\Select\Searchable as Component;
-use TallStackUi\View\Personalizations\Contracts\Personalizable;
-use TallStackUi\View\Personalizations\PersonalizationResource;
-
-/**
- * @internal This class is not meant to be used directly.
- */
-class Searchable extends PersonalizationResource implements Personalizable
-{
-    protected function component(): string
-    {
-        return Component::class;
-    }
-}
