+++ src/View/Personalizations/Components/Select/Select.php
@@ -2,7 +2,7 @@
 
 namespace TallStackUi\View\Personalizations\Components\Select;
 
-use TallStackUi\View\Components\Select\Select as Component;
+use TallStackUi\View\Components\Select\Native as Component;
 use TallStackUi\View\Personalizations\Contracts\Personalizable;
 use TallStackUi\View\Personalizations\PersonalizationResource;
