+++ src/View/Personalizations/Personalization.php
@@ -26,9 +26,7 @@ use TallStackUi\View\Personalizations\Components\Hint;
 use TallStackUi\View\Personalizations\Components\Interactions\Dialog;
 use TallStackUi\View\Personalizations\Components\Interactions\Toast;
 use TallStackUi\View\Personalizations\Components\Modal;
-use TallStackUi\View\Personalizations\Components\Select\Searchable as SelectSearchable;
 use TallStackUi\View\Personalizations\Components\Select\Select;
-use TallStackUi\View\Personalizations\Components\Select\Style as SelectStyle;
 use TallStackUi\View\Personalizations\Components\Select\Styled as SelectStyled;
 use TallStackUi\View\Personalizations\Components\Tabs\Items as TabItems;
 use TallStackUi\View\Personalizations\Components\Tabs\Tab;
@@ -64,10 +62,8 @@ class Personalization
         'tallstack-ui::personalizations.form.toggle' => Toggle::class,
         'tallstack-ui::personalizations.hint' => Hint::class,
         'tallstack-ui::personalizations.modal' => Modal::class,
-        'tallstack-ui::personalizations.select' => Select::class,
-        'tallstack-ui::personalizations.select.searchable' => SelectSearchable::class,
+        'tallstack-ui::personalizations.select.native' => Select::class,
         'tallstack-ui::personalizations.select.styled' => SelectStyled::class,
-        'tallstack-ui::personalizations.select.style' => SelectStyle::class,
         'tallstack-ui::personalizations.tab' => Tab::class,
         'tallstack-ui::personalizations.tab.items' => TabItems::class,
         'tallstack-ui::personalizations.toast' => Toast::class,
@@ -198,15 +194,13 @@ class Personalization
         return app($this->component(Modal::class));
     }
 
-    public function select(string $component = null): Select|SelectSearchable|SelectStyled|SelectStyle
+    public function select(string $component = null): Select|SelectStyled
     {
-        $component ??= 'select';
+        $component ??= 'native';
 
         $class = match ($component) {
-            'select' => Select::class,
-            'searchable' => SelectSearchable::class,
+            'native' => Select::class,
             'styled' => SelectStyled::class,
-            'style' => SelectStyle::class,
             default => $component,
         };
