+++ src/config.php
@@ -70,10 +70,8 @@ return [
         'modal' => Components\Modal::class,
         'password' => Components\Form\Password::class,
         'radio' => Components\Form\Radio::class,
-        'select.native' => Components\Select\Select::class,
-        'select.searchable' => Components\Select\Searchable::class,
+        'select.native' => Components\Select\Native::class,
         'select.styled' => Components\Select\Styled::class,
-        'select.style' => Components\Select\Style::class,
         'tab' => Components\Tab\Tab::class,
         'tab.items' => Components\Tab\Items::class,
         'textarea' => Components\Form\Textarea::class,
